Reduce the core of m8t to a data processing pipeline.
Previously we had a load phase, followed by a render phase which understood
layouts and how to default various bits of page data.

Now, m8t just iteratively processes data with a set of processors. A processor
is anything that has a process method that takes in page data and either returns

1. nothing if it does not know how/want to process the given data, or
2. a new piece of data.

Although this model simplifies everything to one concept, there's implicitly
three kinds of processors.

First, we have loaders, that essentially prime the data processing pipeline with
filenames and the base data from files. For example, we may have loaders that
load contents from markdown files, or loaders that can take a text file and
produce multiple data items for each paragraph in the file.

Second, we have transformers, which take a piece of data and tweak it.
Typically, transformers don't touch existing keys in the data object, but add
new pieces of data, such as the reading time for the content key in the data or
adding a default title based on the filename.

Finally, we have renderers, which finalize data by combining it all into a piece
of content that could be served to the end user. This could be something that
takes a React element and produces HTML, or a Javascript bundler to produce
JavaScript files and chunks for their dependencies.