Move to a more explicit pipeline specification.
Previously, we had a list of processors, and every time a processor processed
some datum, that datum would then be put through all processors again and again,
until eventually it came out the other end. This works fine, but it requires
every processor to think about "have I already processed this data?", and
generally is a bit harder to reason about.

We now view a pipeline as a series of stages, each stage getting all data to
process. Once processed, all data moves to the next stage. If data _does_ need
to be processed by the entire pipeline, there's an escape hatch: they can add
the `[reprocess]: true` to the data they emit.

Stages can do whatever they want with the incoming data. For example, they could

- pass data through untouched,
- transform a piece of data, or
- add new data.

In the previous world, this is mostly how things worked, but there were three
challenges moving to this new model.

First, sometimes we did need to reprocess data. For example, in my personal
website there was an `icons.tsx` that emitted (its content function was a
generator) many React elements that render to SVG images. There weren't any good
indicators from that emitted data to detect that the data needed to be
reprocessed, so we now need to explicitly say "yes, put me at the beginning of
the pipeline".

Second, layouts sort of needed to be reprocessed to, but only by some of the
processors. On top of that, it's not new data, so we needed to properly track
the lineage of the data and not clobber it with the layout file. To solve this
problem, we added an option to the layout transformer so that it can take a
pipeline itself that will be used to process the layout file.

Third, search can be tricky. Layouts have searches, and when we perform them we
want to defer getting the data. Previously we generated a new object instance
every time, but that was okay because we had a mechanism to defer searches until
the pipeline was busy. That's no longer the case, but the data returned from
searches can still evolve before the rendering pass. This was the introduction
of `Datum`, which also tracks the lineage. It's a bit more verbose to access the
data (e.g., `datum.get("blah")` vs `datum.blah`) but I think having this stable
identity will pay its dividends in the long term. It also has nice names for
things that will likely help anyone when implementing their own custom stages.

There's definitely still some work to be done here, but overall I believe this
will be a win. For example, the static javascript processor is far less janky,
and can operate independently from the thing running the pipeline (previously,
that thing would have to ask for and emit the chunks itself).