Fix bug where reloader wasn't reloading in certain situations.
We debounce the reload, but a debounced function only receives the last set of
arguments given to it. This is fine in many situations, like a user typing in
the frontend, but for us we need all of the arguments ever given to the reload
function to see if at least one of them is a path that requires reload.