Fix dev server static file serving bug.
We were trying to be clever in 0.7.2 by having a `finally` block that would
check to see if the route handler dealt with the response. If not, we had it
finish the response with "no content".

The bug is that the route handler wasn't expected to always complete the request
after returning. This was the case with static file serving, where we pipe a
read stream to the response, so the ending will be offloaded to the event loop.

Instead, we use `setTimeout` to write a response if nothing already written.