Improve pipeline processing for reprocessed data.
Previous we reprocessed data concurrently with the stage that produced
reprocessed data. We now wait for the reprocessed data to catch up with the
stage that produced it, which allows us to combine this data into one big data
array that can be processed by future stages.

This allows future stages to process _all_ data at once. Most stages don't care
about this, but some do. For example, the `TypesProcessor` will want to produce
a since `.d.ts` file, so seeing all possible data at once is ideal. Similarly,
stages that produce a search index may want to do that one-shot and write an
index.